exercicio 46 refeito codigo - treino
